Main document for destination weights
This is the main document that has the key calculations for generating the weights as per the MATSim destinations input file to model accessibility. The document contains selection of the main Points of Interest classes, the number of employment in each industry sector, and average number of employees per sector and industry class. The goal is to define the attractiveness of each of the POI destinations, based on [1] the size of each of the POI building footprint and [2] the number of people employed.